feat: add controller scanner
